first iteration of the wrapper script
The wrapper script is described in the readme. It's how users and
Jenkins will interact with the toolkit. It has to handle every command
that this project will support. This shows rudimentary install
behavior as well as how the get command works. Other commands are
passed through to the internal script. Most of the install logic will
also live in the internal script. That script is more complex and left
for another PR.