straightens out the testing action
The action was failing because it was trying to shellcheck everything
in the tests directory, which includes another directory. This makes
it test only files in both internal and tests, assuming that all files
will be shell scripts.

If and when non-shell scripts get added, figure out a different way.

This also adds tests to ensure the project structure stays predictable
so that scripts and tests won't evade the notice of shellcheck or bats
in the future by being added where they're not looking.