use if/then where appropriate
My first draft of the internal script used && and || for lots of
decision making and didn't do so very safely. shellcheck caught that,
and I broke it down some, but if/then reads more naturally in many of
these situations, so this changes many of those conditions to use
if/then.

My main goal with this change is readability.