add an update check to the wrapper script
In order to encourage users to keep up to date, this adds an update
check to the wrapper script to let them know if a new version of that
script is available.

Update checks for internal files can be handled internally.

Also fixes a test that somehow got committed failing because of
changes to the output a while back.