tidying up rax-docs script
Just a few minor changes to clean up. No functional change. The
SOURCE_DIR variable to change cloning behavior didn't work out very
well, so it's removed here. The problem is that even if you copy your
local repo dir into the working directory for tests, tests would still
need to check out some version of it, which would change the code
being run and make the tests be testing code they weren't necessarily
written for. It's better to just stub out git in the tests and check
that the right commands are happening.