add first integration test
As I continue testing, it becomes apparent that it'll be useful to
divide unit (fast) tests from integration (maybe slow) tests. Building
the docker image is one such test. It can take quite a while, and it
also has the side effect of overwriting your local image, so it's not
one you want to include in the "run all the time" test suite.

This adds initial docker image tests as the first integration test.