unit test tidying
Adds a few unit tests, cleans up a little, and moves the test fixtures
into a single directory. It's easier that way so that when you're
tab-completing test file names, you don't run into the fixture dirs.

This establishes a convention (enforced by testing) that the fixture
dir for any given test file is in the fixtures dir and named the same
as the test file.