adds image versioning
If a user is running two or more different docs projects that use
different toolkit versions, the docker images for them would conflict
because it isn't tagged with a version. This fixes that to tag the
image with the local toolkit version. Now building once will suffice
for all projects using that version of the toolkit, and you're free to
use different versions on different projects, too.