changes from beta test
The first beta test revealed a few things to fix.

The bigger issues were around cloning the docs project and the
toolkit.  We need to have the repo url of the docs project put into
the Jenkinsfile by the install process because the config isn't
present to load until after we've already cloned the project. For the
toolkit, we don't know how to get Jenkins to pass credentials through
to the rax-docs script to do a 'get' of the toolkit repo, so we clone
it from the public url instead.