check out version on wrapper install
All this time, the wrapper has been leaving the version checkout up to
the internal script, which means the wrapper checks out master and
runs *that* version of the internal script, which is clearly
wrong. This moves the initial version checkout to the wrapper so things
will work as expected.

The internal script still has lots of other version management things
to do. I think they'll work as expected still.