update toolkit and jenkins build to python 3
Something just changed on the server hosting files.pythonhosted.org,
where pip seems to be downloading packages from. It's using SNI and/or
SAN differently, and our current python 2 isn't handling it
properly. It looks like we're upgrading to python 3 now.

All the changes to requirements.txt were forced on us by the python
version.  It seems that newer versions worked in python 2.7, but
something is enforcing lower versions of many libraries on python 3.3.