restore "install latest" to original behavior
I recently realized that the wrapper needs to check out the right
toolkit version during install so that it's running the correct
version of the internal scripts at install time. I added the "default
to master" checkout at that time, but it's more correct to just clone
without checking out and defer to the "install latest" logic in the
internal script.