jenkins uses the rax-docs requirements.txt
Due to a few minor oversights, the setup_jenkins function that gets
the build env ready on jenkins was using a project-local
requirements.txt. This was left over in the docs project I tested the
function out with because the "old files cleanup" wasn't running on it
properly. This switches jenkins to use the toolkit's requirements.txt
instead.