another nonexistent ref in publish
For the same reason as the previous change, `git push origin gh-pages`
fails because it's a synonym for `git push origin gh-pages:gh-pages`,
and as the last change established, `gh-pages` doesn't exist
locally. Since it exists on the remote, we can simply push HEAD to it,
and all should be well.