action to require labels on PRs
Since the changelog and release notes are built based on labels, and
no label means a PR doesn't show up, we need to encourage label
use. This requires one of the labels recognized by the changelog
action, or "internal", which I hope will serve to mark a PR as
something that doesn't need to be in the release notes.